+++ src/app/blog/(posts)/first/page.mdx
@@ -1,5 +1,5 @@
 export const metadata = {
-  title: 'Hello world',
+  title: 'first',
   description: 'Hello World Page',
   publishDate: '2024-01-01T00:00:00Z',
   posterImage:
