+++ src/types/types.ts
@@ -24,3 +24,12 @@ export interface ImageInfo extends BlurImageSrc {
   width: number;
   height: number;
 }
+
+export interface Post {
+  slug: string;
+  title: string;
+  description: string;
+  publishDate: string;
+  posterImage: string;
+  categories: string[];
+}
